use Syncable
1. replace SyncSourcing
2. update example code
3. make Syncable public
